v1.0.3.9005
updated getUSGSflow function to fill in missing flow values at the beginning and ending of time series. (will need additional work to migrate to only fill in values up to a gap of max.fill, but then also move the max.fill argument so that it is exposed to user selection).